+++ lib/store/helpers/move/index.ts
@@ -1,4 +1,6 @@
 import { moveUp } from "./moveUp";
 import { moveDown } from "./moveDown";
+import { subMoveUp } from "./subMoveUp";
+import { subMoveDown } from "./subMoveDown";
 
-export { moveUp, moveDown };
+export { moveUp, moveDown, subMoveUp, subMoveDown };
